requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.2.0
boto3==1.34.88
botocore==1.34.88
celery==5.3.6
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorlog==6.8.2
constantly==23.10.4
cryptography==42.0.7
daphne==4.1.2
dj-database-url==0.5.0
Django==5.0.4
django-cors-headers==4.3.1
django-storages==1.10.1
djangorestframework==3.14.0
djangorestframework-api-key==2.3.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.7.1
et-xmlfile==1.1.0
gunicorn==22.0.0
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
inflection==0.5.1
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kombu==5.3.4
msgpack==1.0.8
openpyxl==3.1.2
packaging==23.2
pillow==10.2.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.1.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
referencing==0.35.1
rpds-py==0.20.0
s3transfer==0.10.1
service-identity==24.1.0
setuptools==69.5.1
six==1.16.0
sqlparse==0.4.4
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==1.26.18
vine==5.1.0
wcwidth==0.2.12
whitenoise==6.6.0
zope.interface==6.3
